Extraer modal de éxito de mark.js
Quinto paso de la descomposición de mark.js: showSuccessModal() y
closeModalHandler() (renombrada closeSuccessModal) se mueven a
mark/success-modal.js. A diferencia del modal de error, este no
comparte estado (errorModalVisible, previousActiveElement) con el
resto del flujo de identificación/cámara — solo necesita un callback
onClose (returnToSplash) que mark.js pasa en sus 2 puntos de llamada.
Se elimina modalListenerAdded de mark.js, ya no lo usa nada más.
Refactor puro, sin cambios de comportamiento.
